+++ handbook/src/pages/index.js
@@ -131,7 +131,7 @@ function Gitee() {
             className={"furion-log-jiao" + (isDarkTheme ? " dark" : "")}
           ></div>
           <div className="furion-log-number">
-            <div style={{ color: "#723cff" }}>3450</div>
+            <div style={{ color: "#723cff" }}>3500</div>
             <span className={isDarkTheme ? " dark" : ""}>Stars</span>
           </div>
         </div>
@@ -155,7 +155,7 @@ function Gitee() {
             className={"furion-log-jiao" + (isDarkTheme ? " dark" : "")}
           ></div>
           <div className="furion-log-number">
-            <div style={{ color: "#1fd898" }}>111,792</div>
+            <div style={{ color: "#1fd898" }}>112,458</div>
             <span className={isDarkTheme ? " dark" : ""}>Downloads</span>
           </div>
         </div>